Refactor POS RBAC and add table reservations

+++ apps/orders/views/order_item_detail.py
@@ -4,7 +4,12 @@ from rest_framework import generics, permissions
 from apps.orders.models import OrderItem
 from apps.orders.serializers import OrderItemSerializer
 from apps.orders.services import OrderStateService
-from common.api.permissions import EndpointRBACPermission
+from common.api.permissions import (
+    EndpointRBACPermission,
+    POS_TABLES_MANAGE_PERMISSION,
+    POS_TAKEAWAY_MENU_VIEW_PERMISSION,
+    require_any_permission_code,
+)
 from common.api.scopes import get_request_restaurant
 
 
@@ -19,6 +24,12 @@ class OrderItemDetailView(generics.RetrieveUpdateDestroyAPIView):
 
     @transaction.atomic
     def perform_update(self, serializer):
+        required_permission = (
+            POS_TABLES_MANAGE_PERMISSION
+            if serializer.instance.order.table_session_id
+            else POS_TAKEAWAY_MENU_VIEW_PERMISSION
+        )
+        require_any_permission_code(self.request.user, required_permission)
         self.state_service_class().ensure_order_mutable(order=serializer.instance.order)
         instance = serializer.save()
         self.state_service_class().sync_after_items_changed(order=instance.order)
@@ -26,6 +37,8 @@ class OrderItemDetailView(generics.RetrieveUpdateDestroyAPIView):
     @transaction.atomic
     def perform_destroy(self, instance):
         order = instance.order
+        required_permission = POS_TABLES_MANAGE_PERMISSION if order.table_session_id else POS_TAKEAWAY_MENU_VIEW_PERMISSION
+        require_any_permission_code(self.request.user, required_permission)
         self.state_service_class().ensure_order_mutable(order=order)
         instance.delete()
         self.state_service_class().sync_after_items_changed(order=order)
